+++ workspace/src/main/java/interfaces/ITSPDisplayer.java
@@ -9,7 +9,7 @@ public interface ITSPDisplayer {
     void displayNodes(short nbNodes);
     void clearNodes();
 
-    void displaySolution(int nbNodes, Short[] solution, long dist, boolean best);
+    void displaySolution(int nbNodes, short[] solution, double dist, boolean best);
     void clearSolutions();
 
     void addListener(ITSPDisplayerCallback l);
